+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0](https://github.com/99stealth/aws-profile-manager/tree/v0.4.0) - 2022-05-15
 ### Added
 - Functionality that allows to rotate aws access keys
+- `--verson` and `--quiet` root options
 ### Changed
 - Function `choose_profile_to_remove` to `choose_profile` in `profile-remove` command
 ### Removed